move policy generators into their respective folders
